reorganize repo with new classification
