+++ legacy/classification/requirements.txt
@@ -1,0 +1,11 @@
+--extra-index-url https://download.pytorch.org/whl/cu102
+torch==1.11.0+cu102
+torchvision==0.12.0+cu102
+tensorboard
+numpy>=1.21.6
+onnx
+onnxruntime
+scikit-learn
+seaborn
+pandas
+matplotlib
